requirements.txt
dbt-core~=1.5.2
dbt-postgres~=1.5.2
sqlfluff==2.3.1
aws-cdk-lib==2.86.0
constructs>=10.0.0,<11.0.0
